Refactor assign
