+++ src/cdb/db.rs
@@ -124,77 +124,6 @@ impl IndexMut<ClauseIndex> for ClauseDB {
     }
 }
 
-// impl Index<&ClauseId> for ClauseDB {
-//     type Output = Clause;
-//     #[inline]
-//     fn index(&self, cid: &ClauseId) -> &Clause {
-//         let i = NonZeroU32::get(cid.ordinal) as usize;
-//         if cfg!(feature = "unsafe_access") {
-//             unsafe { self.clause.get_unchecked(i) }
-//         } else {
-//             &self.clause[i]
-//         }
-//     }
-// }
-
-// impl IndexMut<&ClauseId> for ClauseDB {
-//     #[inline]
-//     fn index_mut(&mut self, cid: &ClauseId) -> &mut Clause {
-//         let i = NonZeroU32::get(cid.ordinal) as usize;
-//         if cfg!(feature = "unsafe_access") {
-//             unsafe { self.clause.get_unchecked_mut(i) }
-//         } else {
-//             &mut self.clause[i]
-//         }
-//     }
-// }
-
-// impl Index<Range<usize>> for ClauseDB {
-//     type Output = [Clause];
-//     #[inline]
-//     fn index(&self, r: Range<usize>) -> &[Clause] {
-//         if cfg!(feature = "unsafe_access") {
-//             unsafe { self.clause.get_unchecked(r) }
-//         } else {
-//             &self.clause[r]
-//         }
-//     }
-// }
-
-// impl Index<RangeFrom<usize>> for ClauseDB {
-//     type Output = [Clause];
-//     #[inline]
-//     fn index(&self, r: RangeFrom<usize>) -> &[Clause] {
-//         if cfg!(feature = "unsafe_access") {
-//             unsafe { self.clause.get_unchecked(r) }
-//         } else {
-//             &self.clause[r]
-//         }
-//     }
-// }
-
-// impl IndexMut<Range<usize>> for ClauseDB {
-//     #[inline]
-//     fn index_mut(&mut self, r: Range<usize>) -> &mut [Clause] {
-//         if cfg!(feature = "unsafe_access") {
-//             unsafe { self.clause.get_unchecked_mut(r) }
-//         } else {
-//             &mut self.clause[r]
-//         }
-//     }
-// }
-
-// impl IndexMut<RangeFrom<usize>> for ClauseDB {
-//     #[inline]
-//     fn index_mut(&mut self, r: RangeFrom<usize>) -> &mut [Clause] {
-//         if cfg!(feature = "unsafe_access") {
-//             unsafe { self.clause.get_unchecked_mut(r) }
-//         } else {
-//             &mut self.clause[r]
-//         }
-//     }
-// }
-
 impl Instantiate for ClauseDB {
     fn instantiate(config: &Config, cnf: &CNFDescription) -> ClauseDB {
         let nv = cnf.num_of_variables;
